row filter tests and rowFilters.ts
