+++ web/src/features/moreCast2/components/ForecastSummaryDataGrid.tsx
@@ -9,7 +9,7 @@ import { DataGridColumns } from 'features/moreCast2/components/DataGridColumns'
 import { storeUserEditedRows, getSimulatedIndices } from 'features/moreCast2/slices/dataSlice'
 import { AppDispatch } from 'app/store'
 import { useDispatch } from 'react-redux'
-import { validActualOrForecastPredicate, validForecastPredicate } from 'features/moreCast2/util'
+import { filterRowsForSimulationFromEdited } from 'features/moreCast2/rowFilters'
 
 const PREFIX = 'ForecastSummaryDataGrid'
 
@@ -35,25 +35,6 @@ interface ForecastSummaryDataGridProps {
   updateColumnWithModel: (modelType: ModelType, colDef: GridColDef) => void
   handleColumnHeaderClick: GridEventListener<'columnHeaderClick'>
   handleClose: () => void
-}
-
-export const filterRowsForSimulationFromEdited = (
-  editedRow: MoreCast2Row,
-  allRows: MoreCast2Row[]
-): MoreCast2Row[] | undefined => {
-  if (validForecastPredicate(editedRow)) {
-    const validRowsForStation = allRows.filter(
-      row => row.stationCode === editedRow.stationCode && validActualOrForecastPredicate(row)
-    )
-
-    const yesterday = editedRow.forDate.minus({ days: 1 })
-    const yesterdayRow = validRowsForStation.find(row => row.forDate.toISODate() === yesterday.toISODate())
-
-    if (yesterdayRow) {
-      const rowsForSimulation = validRowsForStation.filter(row => row.forDate >= yesterday)
-      return rowsForSimulation
-    } else return undefined
-  }
 }
 
 const ForecastSummaryDataGrid = ({
